Apply suggestions and refactor
