+++ services/blockchain-indexer/methods/dataService/modules/token.js
@@ -16,7 +16,7 @@
 const {
 	getTokens,
 	getTokensSummary,
-} = require('../controllers/tokens');
+} = require('../controllers/token');
 
 module.exports = [
 	{
