+++ services/blockchain-indexer/shared/dataService/business/peers.js
@@ -15,47 +15,8 @@
  */
 const { requestConnector } = require('../../utils/request');
 
-const keyValueTable = require('../../database/mysqlKVStore');
-
-const { getPosTokenID } = require('./pos/constants');
-
-const { KEY_VALUE_TABLE_KEYS } = require('../../constants');
-
-const getTotalStaked = async () => {
-	const result = await keyValueTable.getPattern(
-		KEY_VALUE_TABLE_KEYS.TOTAL_STAKED_PREFIX,
-	);
-	const [row] = result;
-
-	let totalStaked = null;
-	if (row && row.value) totalStaked = row.value.toString();
-
-	return {
-		amount: totalStaked,
-		tokenID: await getPosTokenID(),
-	};
-};
-
-const getTotalLocked = async () => {
-	const lockAmountsInfo = await keyValueTable.getPattern(
-		KEY_VALUE_TABLE_KEYS.TOTAL_LOCKED_PREFIX,
-	);
-
-	const totalLockedResponse = lockAmountsInfo.map(({ key, value }) => {
-		const tokenID = key.split('_').pop();
-		return {
-			tokenID,
-			amount: value.toString(),
-		};
-	});
-
-	return totalLockedResponse;
-};
-
 const getPeersStatistics = async () => {
 	const response = await requestConnector('getPeersStatistics');
-	response.totalLocked = await getTotalLocked();
-	response.totalStaked = await getTotalStaked();
 	return {
 		data: response,
 		meta: {},
