+++ services/gateway/sources/version3/mappings/amountEntry.js
